vm: Move VMOpts to types, add index.ts
