+++ packages/vm/src/index.ts
@@ -1,0 +1,2 @@
+export { VM as default } from './vm'
+export { VMOpts } from './types'
