+++ packages/vm/src/vm.ts
@@ -1,4 +1,4 @@
-import { Account, Address, BigIntLike, toType, TypeOutput } from '@ethereumjs/util'
+import { Account, Address, toType, TypeOutput } from '@ethereumjs/util'
 import Blockchain from '@ethereumjs/blockchain'
 import Common, { Chain, Hardfork } from '@ethereumjs/common'
 import { StateManager, DefaultStateManager } from '@ethereumjs/statemanager'
@@ -7,118 +7,10 @@ import { default as runBlock, RunBlockOpts, RunBlockResult } from './runBlock'
 import { default as buildBlock, BuildBlockOpts, BlockBuilder } from './buildBlock'
 import AsyncEventEmitter = require('async-eventemitter')
 import { promisify } from 'util'
-import { VMEvents } from './types'
+import { VMEvents, VMOpts } from './types'
 
 import EVM, { getActivePrecompiles, EEIInterface, EVMInterface } from '@ethereumjs/evm'
 import EEI from './eei/eei'
-
-/**
- * Options for instantiating a {@link VM}.
- */
-export interface VMOpts {
-  /**
-   * Use a {@link Common} instance
-   * if you want to change the chain setup.
-   *
-   * ### Possible Values
-   *
-   * - `chain`: all chains supported by `Common` or a custom chain
-   * - `hardfork`: `mainnet` hardforks up to the `London` hardfork
-   * - `eips`: `2537` (usage e.g. `eips: [ 2537, ]`)
-   *
-   * ### Supported EIPs
-   *
-   * - [EIP-1153](https://eips.ethereum.org/EIPS/eip-1153) - Transient Storage Opcodes (`experimental`)
-   * - [EIP-1559](https://eips.ethereum.org/EIPS/eip-1559) - EIP-1559 Fee Market
-   * - [EIP-2315](https://eips.ethereum.org/EIPS/eip-2315) - VM simple subroutines (`experimental`)
-   * - [EIP-2537](https://eips.ethereum.org/EIPS/eip-2537) - BLS12-381 precompiles (`experimental`)
-   * - [EIP-2565](https://eips.ethereum.org/EIPS/eip-2565) - ModExp Gas Cost
-   * - [EIP-2718](https://eips.ethereum.org/EIPS/eip-2718) - Typed Transactions
-   * - [EIP-2929](https://eips.ethereum.org/EIPS/eip-2929) - Gas cost increases for state access opcodes
-   * - [EIP-2930](https://eips.ethereum.org/EIPS/eip-2930) - Access List Transaction Type
-   * - [EIP-3074](https://eips.ethereum.org/EIPS/eip-3074) - AUTH and AUTHCALL opcodes
-   * - [EIP-3198](https://eips.ethereum.org/EIPS/eip-3198) - BASEFEE opcode
-   * - [EIP-3529](https://eips.ethereum.org/EIPS/eip-3529) - Reduction in refunds
-   * - [EIP-3540](https://eips.ethereum.org/EIPS/eip-3541) - EVM Object Format (EOF) v1 (`experimental`)
-   * - [EIP-3541](https://eips.ethereum.org/EIPS/eip-3541) - Reject new contracts starting with the 0xEF byte
-   *   [EIP-3651](https://eips.ethereum.org/EIPS/eip-3651) - Warm COINBASE (`experimental`)
-   * - [EIP-3670](https://eips.ethereum.org/EIPS/eip-3670) - EOF - Code Validation (`experimental`)
-   * - [EIP-3855](https://eips.ethereum.org/EIPS/eip-3855) - PUSH0 instruction (`experimental`)
-   * - [EIP-3860](https://eips.ethereum.org/EIPS/eip-3860) - Limit and meter initcode (`experimental`)
-   * - [EIP-4399](https://eips.ethereum.org/EIPS/eip-4399) - Supplant DIFFICULTY opcode with PREVRANDAO (Merge) (`experimental`)
-   *
-   * *Annotations:*
-   *
-   * - `experimental`: behaviour can change on patch versions
-   *
-   * ### Default Setup
-   *
-   * Default setup if no `Common` instance is provided:
-   *
-   * - `chain`: `mainnet`
-   * - `hardfork`: `london`
-   * - `eips`: `[]`
-   */
-  common?: Common
-  /**
-   * A {@link StateManager} instance to use as the state store
-   */
-  stateManager?: StateManager
-  /**
-   * A {@link Blockchain} object for storing/retrieving blocks
-   */
-  blockchain?: Blockchain
-  /**
-   * If true, create entries in the state tree for the precompiled contracts, saving some gas the
-   * first time each of them is called.
-   *
-   * If this parameter is false, each call to each of them has to pay an extra 25000 gas
-   * for creating the account. If the account is still empty after this call, it will be deleted,
-   * such that this extra cost has to be paid again.
-   *
-   * Setting this to true has the effect of precompiled contracts' gas costs matching mainnet's from
-   * the very first call, which is intended for testing networks.
-   *
-   * Default: `false`
-   */
-  activatePrecompiles?: boolean
-  /**
-   * If true, the state of the VM will add the genesis state given by {@link Blockchain.genesisState} to a newly
-   * created state manager instance. Note that if stateManager option is also passed as argument
-   * this flag won't have any effect.
-   *
-   * Default: `false`
-   */
-  activateGenesisState?: boolean
-
-  /**
-   * Select hardfork based upon block number. This automatically switches to the right hard fork based upon the block number.
-   *
-   * Default: `false`
-   */
-  hardforkByBlockNumber?: boolean
-  /**
-   * Select the HF by total difficulty (Merge HF)
-   *
-   * This option is a superset of `hardforkByBlockNumber` (so only use one of both options)
-   * and determines the HF by both the block number and the TD.
-   *
-   * Since the TD is only a threshold the block number will in doubt take precedence (imagine
-   * e.g. both Merge and Shanghai HF blocks set and the block number from the block provided
-   * pointing to a Shanghai block: this will lead to set the HF as Shanghai and not the Merge).
-   */
-  hardforkByTD?: BigIntLike
-
-  /**
-   * Use a custom EEI for the EVM. If this is not present, use the default EEI.
-   */
-  eei?: EEIInterface
-
-  /**
-   * Use a custom EVM to run Messages on. If this is not present, use the default EVM.
-   */
-  evm?: EVMInterface
-}
 
 /**
  * Execution engine which can be used to run a blockchain, individual
@@ -126,7 +18,7 @@ export interface VMOpts {
  *
  * This class is an AsyncEventEmitter, please consult the README to learn how to use it.
  */
-export default class VM extends AsyncEventEmitter<VMEvents> {
+export class VM extends AsyncEventEmitter<VMEvents> {
   /**
    * The StateManager used by the VM
    */
